+++ commands/default_cmdsets.py
@@ -48,6 +48,7 @@ from features.world_bosses.commands import WorldBossCmdSet
 from features.runes.commands import RunasCmdSet
 from features.arena.commands import ArenaCmdSet
 from features.housing.commands import HousingCmdSet
+from features.bestiary.commands import BestiaryCmdSet
 from commands.general_commands import GeneralCmdSet
 
 
@@ -133,6 +134,8 @@ class CharacterCmdSet(default_cmds.CharacterCmdSet):
         self.add(ArenaCmdSet)
         # --- Vivienda ---
         self.add(HousingCmdSet)
+        # --- Bestiario ---
+        self.add(BestiaryCmdSet)
         # --- General ---
         self.add(GeneralCmdSet)